+++ Z/Data/Builder/Numeric.hs
@@ -709,19 +709,19 @@ foreign import ccall unsafe "static grisu3" c_grisu3
 -- | Decimal encoding of a 'Double', note grisu only handles strictly positive finite numbers.
 grisu3 :: Double -> ([Int], Int)
 {-# INLINE grisu3 #-}
-grisu3 d = snd . unsafePerformIO $
-    allocMutablePrimArrayUnsafe @Word8 GRISU3_DOUBLE_BUF_LEN $ \ pBuf -> do
-        (len, (e, success)) <- allocPrimUnsafe $ \ pLen ->
-            allocPrimUnsafe $ \ pE ->
-                c_grisu3 (realToFrac d) pBuf pLen pE
-        if success == 0 -- grisu3 fail
-        then pure (floatToDigits 10 d)
-        else do
-            buf <- forM [0..len-1] $ \ i -> do
-                w8 <- readByteArray (MutableByteArray pBuf) i :: IO Word8
-                pure $! fromIntegral w8
-            let !e' = e + len
-            pure (buf, e')
+grisu3 d = unsafePerformIO $ do
+    (MutableByteArray pBuf) <- newByteArray GRISU3_DOUBLE_BUF_LEN
+    (len, (e, success)) <- allocPrimUnsafe $ \ pLen ->
+        allocPrimUnsafe $ \ pE ->
+            c_grisu3 (realToFrac d) pBuf pLen pE
+    if success == 0 -- grisu3 fail
+    then pure (floatToDigits 10 d)
+    else do
+        buf <- forM [0..len-1] $ \ i -> do
+            w8 <- readByteArray (MutableByteArray pBuf) i :: IO Word8
+            pure $! fromIntegral w8
+        let !e' = e + len
+        pure (buf, e')
 
 foreign import ccall unsafe "static grisu3_sp" c_grisu3_sp
     :: Float
@@ -733,19 +733,19 @@ foreign import ccall unsafe "static grisu3_sp" c_grisu3_sp
 -- | Decimal encoding of a 'Float', note grisu3_sp only handles strictly positive finite numbers.
 grisu3_sp :: Float -> ([Int], Int)
 {-# INLINE grisu3_sp #-}
-grisu3_sp d = snd . unsafePerformIO $
-    allocMutablePrimArrayUnsafe @Word8 GRISU3_SINGLE_BUF_LEN $ \ pBuf -> do
-        (len, (e, success)) <- allocPrimUnsafe $ \ pLen ->
-            allocPrimUnsafe $ \ pE ->
-                c_grisu3_sp (realToFrac d) pBuf pLen pE
-        if success == 0 -- grisu3 fail
-        then pure (floatToDigits 10 d)
-        else do
-            buf <- forM [0..len-1] $ \ i -> do
-                w8 <- readByteArray (MutableByteArray pBuf) i :: IO Word8
-                pure $! fromIntegral w8
-            let !e' = e + len
-            pure (buf, e')
+grisu3_sp d = unsafePerformIO $ do
+    (MutableByteArray pBuf) <- newByteArray GRISU3_SINGLE_BUF_LEN
+    (len, (e, success)) <- allocPrimUnsafe $ \ pLen ->
+        allocPrimUnsafe $ \ pE ->
+            c_grisu3_sp (realToFrac d) pBuf pLen pE
+    if success == 0 -- grisu3 fail
+    then pure (floatToDigits 10 d)
+    else do
+        buf <- forM [0..len-1] $ \ i -> do
+            w8 <- readByteArray (MutableByteArray pBuf) i :: IO Word8
+            pure $! fromIntegral w8
+        let !e' = e + len
+        pure (buf, e')
 
 --------------------------------------------------------------------------------