+++ Z/Data/Vector.hs
@@ -117,6 +117,7 @@ module Z.Data.Vector (
   , uncons, unsnoc
   , headMaybe, tailMayEmpty
   , lastMaybe, initMayEmpty
+  , head, tail, last, init
   , inits, tails
   , take, drop, takeR, dropR
   , slice